Refactor BucketIndex
