+++ src/bucket/BucketUtils.cpp
@@ -3,10 +3,52 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "bucket/BucketUtils.h"
+#include "bucket/HotArchiveBucket.h"
+#include "bucket/LiveBucket.h"
+#include "ledger/LedgerTypeUtils.h"
+#include "main/Application.h"
+#include "util/types.h"
+#include "xdr/Stellar-ledger-entries.h"
+#include <fmt/format.h>
+#include <medida/counter.h>
 #include <medida/metrics_registry.h>
 
 namespace stellar
 {
+
+std::string
+toString(LedgerEntryTypeAndDurability const type)
+{
+    switch (type)
+    {
+    case LedgerEntryTypeAndDurability::ACCOUNT:
+        return "ACCOUNT";
+    case LedgerEntryTypeAndDurability::TRUSTLINE:
+        return "TRUSTLINE";
+    case LedgerEntryTypeAndDurability::OFFER:
+        return "OFFER";
+    case LedgerEntryTypeAndDurability::DATA:
+        return "DATA";
+    case LedgerEntryTypeAndDurability::CLAIMABLE_BALANCE:
+        return "CLAIMABLE_BALANCE";
+    case LedgerEntryTypeAndDurability::LIQUIDITY_POOL:
+        return "LIQUIDITY_POOL";
+    case LedgerEntryTypeAndDurability::TEMPORARY_CONTRACT_DATA:
+        return "TEMPORARY_CONTRACT_DATA";
+    case LedgerEntryTypeAndDurability::PERSISTENT_CONTRACT_DATA:
+        return "PERSISTENT_CONTRACT_DATA";
+    case LedgerEntryTypeAndDurability::CONTRACT_CODE:
+        return "CONTRACT_CODE";
+    case LedgerEntryTypeAndDurability::CONFIG_SETTING:
+        return "CONFIG_SETTING";
+    case LedgerEntryTypeAndDurability::TTL:
+        return "TTL";
+    default:
+        throw std::runtime_error(
+            fmt::format(FMT_STRING("unknown LedgerEntryTypeAndDurability {:d}"),
+                        static_cast<uint32_t>(type)));
+    }
+}
 
 MergeCounters&
 MergeCounters::operator+=(MergeCounters const& delta)
@@ -150,4 +192,153 @@ EvictionStatistics::submitMetricsAndRestartCycle(uint32_t currLedgerSeq,
     mNumEntriesEvicted = 0;
     mEvictionCycleStartLedger = currLedgerSeq;
 }
+
+template <>
+LedgerEntryTypeAndDurability
+bucketEntryToLedgerEntryAndDurabilityType<LiveBucket>(
+    LiveBucket::EntryT const& be)
+{
+    auto bet = be.type();
+    LedgerKey key;
+    if (bet == INITENTRY || bet == LIVEENTRY)
+    {
+        key = LedgerEntryKey(be.liveEntry());
+    }
+    else if (bet == DEADENTRY)
+    {
+        key = be.deadEntry();
+    }
+    else
+    {
+        throw std::runtime_error("Unexpected Bucket Meta Entry");
+    }
+
+    switch (key.type())
+    {
+    case ACCOUNT:
+        return LedgerEntryTypeAndDurability::ACCOUNT;
+    case TRUSTLINE:
+        return LedgerEntryTypeAndDurability::TRUSTLINE;
+    case OFFER:
+        return LedgerEntryTypeAndDurability::OFFER;
+    case DATA:
+        return LedgerEntryTypeAndDurability::DATA;
+    case CLAIMABLE_BALANCE:
+        return LedgerEntryTypeAndDurability::CLAIMABLE_BALANCE;
+    case LIQUIDITY_POOL:
+        return LedgerEntryTypeAndDurability::LIQUIDITY_POOL;
+    case CONTRACT_DATA:
+        return isTemporaryEntry(key)
+                   ? LedgerEntryTypeAndDurability::TEMPORARY_CONTRACT_DATA
+                   : LedgerEntryTypeAndDurability::PERSISTENT_CONTRACT_DATA;
+    case CONTRACT_CODE:
+        return LedgerEntryTypeAndDurability::CONTRACT_CODE;
+    case CONFIG_SETTING:
+        return LedgerEntryTypeAndDurability::CONFIG_SETTING;
+    case TTL:
+        return LedgerEntryTypeAndDurability::TTL;
+    default:
+        auto label = xdr::xdr_traits<LedgerEntryType>::enum_name(key.type());
+        throw std::runtime_error(
+            fmt::format("Unknown LedgerEntryType {}", label));
+    }
+}
+
+template <>
+LedgerEntryTypeAndDurability
+bucketEntryToLedgerEntryAndDurabilityType<HotArchiveBucket>(
+    HotArchiveBucket::EntryT const& be)
+{
+    auto bet = be.type();
+    LedgerKey key;
+    if (bet == HOT_ARCHIVE_ARCHIVED)
+    {
+        key = LedgerEntryKey(be.archivedEntry());
+    }
+    else if (bet == HOT_ARCHIVE_LIVE || bet == HOT_ARCHIVE_DELETED)
+    {
+        key = be.key();
+    }
+    else
+    {
+        throw std::runtime_error("Unexpected Hot Archive Meta Entry");
+    }
+
+    if (key.type() != CONTRACT_DATA && isTemporaryEntry(key))
+    {
+        throw std::runtime_error("HotArchiveBucketEntry is temporary");
+    }
+
+    switch (key.type())
+    {
+    case CONTRACT_DATA:
+        return LedgerEntryTypeAndDurability::PERSISTENT_CONTRACT_DATA;
+    case CONTRACT_CODE:
+        return LedgerEntryTypeAndDurability::CONTRACT_CODE;
+    default:
+        auto label = xdr::xdr_traits<LedgerEntryType>::enum_name(key.type());
+        throw std::runtime_error(
+            fmt::format("Unexpected LedgerEntryType in HotArchive: {}", label));
+    }
+}
+
+template <>
+bool
+isBucketMetaEntry<HotArchiveBucket>(HotArchiveBucket::EntryT const& be)
+{
+    return be.type() == HOT_ARCHIVE_METAENTRY;
+}
+
+template <>
+bool
+isBucketMetaEntry<LiveBucket>(LiveBucket::EntryT const& be)
+{
+    return be.type() == METAENTRY;
+}
+
+template <class BucketT>
+void
+BucketEntryCounters::count(typename BucketT::EntryT const& be)
+{
+    if (isBucketMetaEntry<BucketT>(be))
+    {
+        // Do not count meta entries.
+        return;
+    }
+    auto ledt = bucketEntryToLedgerEntryAndDurabilityType<BucketT>(be);
+    entryTypeCounts[ledt]++;
+    entryTypeSizes[ledt] += xdr::xdr_size(be);
+}
+
+BucketEntryCounters&
+BucketEntryCounters::operator+=(BucketEntryCounters const& other)
+{
+    for (auto [type, count] : other.entryTypeCounts)
+    {
+        this->entryTypeCounts[type] += count;
+    }
+    for (auto [type, size] : other.entryTypeSizes)
+    {
+        this->entryTypeSizes[type] += size;
+    }
+    return *this;
+}
+
+bool
+BucketEntryCounters::operator==(BucketEntryCounters const& other) const
+{
+    return this->entryTypeCounts == other.entryTypeCounts &&
+           this->entryTypeSizes == other.entryTypeSizes;
+}
+
+bool
+BucketEntryCounters::operator!=(BucketEntryCounters const& other) const
+{
+    return !(*this == other);
+}
+
+template void
+BucketEntryCounters::count<LiveBucket>(LiveBucket::EntryT const& be);
+template void BucketEntryCounters::count<HotArchiveBucket>(
+    HotArchiveBucket::EntryT const& be);
 }
