+++ src/bucket/BucketUtils.h
@@ -4,10 +4,10 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "main/Application.h"
 #include "xdr/Stellar-ledger-entries.h"
 #include <cstdint>
 #include <list>
+#include <map>
 #include <mutex>
 
 namespace medida
@@ -18,6 +18,8 @@ namespace stellar
 {
 
 class Application;
+class LiveBucket;
+class HotArchiveBucket;
 
 #define BUCKET_TYPE_ASSERT(BucketT) \
     static_assert(std::is_same_v<BucketT, LiveBucket> || \
@@ -156,4 +158,46 @@ class EvictionStatistics
     void submitMetricsAndRestartCycle(uint32_t currLedgerSeq,
                                       EvictionCounters& counters);
 };
+
+enum class LedgerEntryTypeAndDurability : uint32_t
+{
+    ACCOUNT = 0,
+    TRUSTLINE = 1,
+    OFFER = 2,
+    DATA = 3,
+    CLAIMABLE_BALANCE = 4,
+    LIQUIDITY_POOL = 5,
+    TEMPORARY_CONTRACT_DATA = 6,
+    PERSISTENT_CONTRACT_DATA = 7,
+    CONTRACT_CODE = 8,
+    CONFIG_SETTING = 9,
+    TTL = 10,
+    NUM_TYPES = 11,
+};
+
+struct BucketEntryCounters
+{
+    std::map<LedgerEntryTypeAndDurability, size_t> entryTypeCounts;
+    std::map<LedgerEntryTypeAndDurability, size_t> entryTypeSizes;
+
+    template <class BucketT> void count(typename BucketT::EntryT const& be);
+    BucketEntryCounters& operator+=(BucketEntryCounters const& other);
+    bool operator==(BucketEntryCounters const& other) const;
+    bool operator!=(BucketEntryCounters const& other) const;
+
+    template <class Archive>
+    void
+    serialize(Archive& ar)
+    {
+        ar(entryTypeCounts, entryTypeSizes);
+    }
+};
+
+template <class BucketT>
+bool isBucketMetaEntry(typename BucketT::EntryT const& be);
+
+template <class BucketT>
+LedgerEntryTypeAndDurability
+bucketEntryToLedgerEntryAndDurabilityType(typename BucketT::EntryT const& be);
+std::string toString(LedgerEntryTypeAndDurability let);
 }
