+++ src/bucket/HotArchiveBucket.cpp
@@ -122,9 +122,9 @@ HotArchiveBucket::getBucketVersion() const
     return it.getMetadata().ledgerVersion;
 }
 
-HotArchiveBucket::HotArchiveBucket(std::string const& filename,
-                                   Hash const& hash,
-                                   std::unique_ptr<BucketIndex const>&& index)
+HotArchiveBucket::HotArchiveBucket(
+    std::string const& filename, Hash const& hash,
+    std::unique_ptr<HotArchiveBucket::IndexT const>&& index)
     : BucketBase(filename, hash, std::move(index))
 {
 }
